requirements.txt
krakenex
pandas
python-dotenv
requests
tabulate
optuna
xgboost
scikit-learn
